+++ R/scoped_na_if.r
@@ -1,5 +1,7 @@
 #' Convert values to NA in multiple columns
 #'
+#' \lifecycle{deprecated}
+#'
 #' The [dplyr::scoped] variants of [na_if()] and [na_if_not()] can be used
 #' directly within pipelines and can modify multiple variables at once.
 #'  * `*_all()` affects every variable
@@ -47,6 +49,7 @@ NULL
 #' @export
 
 na_if_all <- function(.tbl, ...) {
+  lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
   scoped_na_if(
     fun       = dplyr::mutate_all,
     .tbl      = .tbl,
@@ -59,6 +62,7 @@ na_if_all <- function(.tbl, ...) {
 #' @export
 
 na_if_not_all <- function(.tbl, ...) {
+  lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
   scoped_na_if(
     fun       = dplyr::mutate_all,
     .tbl      = .tbl,
@@ -72,6 +76,7 @@ na_if_not_all <- function(.tbl, ...) {
 #' @export
 
 na_if_at <- function(.tbl, .vars, ...) {
+  lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
   scoped_na_if(
     fun       = dplyr::mutate_at,
     .tbl      = .tbl,
@@ -85,6 +90,7 @@ na_if_at <- function(.tbl, .vars, ...) {
 #' @export
 
 na_if_not_at <- function(.tbl, .vars, ...) {
+  lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
   scoped_na_if(
     fun       = dplyr::mutate_at,
     .tbl      = .tbl,
@@ -99,6 +105,7 @@ na_if_not_at <- function(.tbl, .vars, ...) {
 #' @export
 
 na_if_if <- function(.tbl, .predicate, ...) {
+  lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
   scoped_na_if(
     fun        = dplyr::mutate_if,
     .tbl       = .tbl,
@@ -112,6 +119,7 @@ na_if_if <- function(.tbl, .predicate, ...) {
 #' @export
 
 na_if_not_if <- function(.tbl, .predicate, ...) {
+  lifecycle::deprecate_soft("0.6.0", format(sys.call()[1]), "dplyr::across()")
   scoped_na_if(
     fun        = dplyr::mutate_if,
     .tbl       = .tbl,